refactor folder structure
Place source code in `src/`
