+++ src/email/base-template.html
@@ -450,9 +450,7 @@
               >
                 <tr>
                   <td class="content-block">
-                    <span class="apple-link"
-                      >Auth Now Inc, Lofotgata 4b, 0458 Oslo</span
-                    >
+                    <span class="apple-link">Auth Now Inc</span>
                     <br />
                     Don't like these emails?
                     <a href="http://i.imgur.com/CScmqnj.gif">Unsubscribe</a>.
