+++ src/email/email.js
@@ -1,4 +1,4 @@
-const auth = require('./auth');
+const auth = require('../auth');
 
 const EMAIL_DOMAIN = process.env.MAILGUN_DOMAIN;
 const generateVerifyEmailMessage = ({ email, hostDomain }) => {
